Move OOM handling to liballoc and remove the `oom` lang item
This removes the need for `no_std` programs to define that lang item themselves:

Fixes https://github.com/rust-lang/rust/issues/51540
