+++ src/librustc/middle/weak_lang_items.rs
@@ -151,5 +151,4 @@ weak_lang_items! {
     panic_impl,         PanicImplLangItem,          rust_begin_unwind;
     eh_personality,     EhPersonalityLangItem,      rust_eh_personality;
     eh_unwind_resume,   EhUnwindResumeLangItem,     rust_eh_unwind_resume;
-    oom,                OomLangItem,                rust_oom;
 }
